refactor: remove unnecesary folder for uploadStaticAssets
