+++ lib/__tests__/uploadStaticAssets.test.js
@@ -1,12 +1,12 @@
 const path = require("path");
-const uploadStaticAssets = require("..");
-const parseNextConfiguration = require("../../parseNextConfiguration");
-const parsedNextConfigurationFactory = require("../../../utils/test/parsedNextConfigurationFactory");
-const ServerlessPluginBuilder = require("../../../utils/test/ServerlessPluginBuilder");
-const uploadDirToS3Factory = require("../../../utils/s3/upload");
+const uploadStaticAssets = require("../uploadStaticAssets");
+const parseNextConfiguration = require("../parseNextConfiguration");
+const parsedNextConfigurationFactory = require("../../utils/test/parsedNextConfigurationFactory");
+const ServerlessPluginBuilder = require("../../utils/test/ServerlessPluginBuilder");
+const uploadDirToS3Factory = require("../../utils/s3/upload");
 
-jest.mock("../../../utils/s3/upload");
-jest.mock("../../parseNextConfiguration");
+jest.mock("../../utils/s3/upload");
+jest.mock("../parseNextConfiguration");
 
 describe("uploadStaticAssets", () => {
   let uploadDirToS3;
