+++ lib/uploadStaticAssets.js
@@ -1,5 +1,5 @@
 const path = require("path");
-const uploadDirToS3Factory = require("../../utils/s3/upload");
+const uploadDirToS3Factory = require("../utils/s3/upload");
 
 module.exports = function() {
   const uploadDirToS3 = uploadDirToS3Factory(this.providerRequest);
